+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+2.1.1
+-----
+- Added id to crossselling entity
+- Added identity linker crossselling id mapping
+
 2.1.0
 -----
 - Changed protocol version to 7 due to product image masterkill sorting